Added error handling routines to tas

+++ tas/include/tas.h
@@ -2,4 +2,4 @@
 
 #include "CLOptions.h"
 
-int64_t tas(CLOptions const& options);
+int tas(CLOptions const& options);
